\name{input_counts}
\alias{input_counts}
\docType{data}
\title{
input data sampled from Hydra data
}
\description{
This is an example input data which is generated by randomly sampling 10000 cells from Hydra data. 
}
\usage{data("input_counts")}
\format{
  The format is:
Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
  ..@ i       : int [1:22443960] 51 62 83 112 122 124 129 130 159 161 ...
  ..@ p       : int [1:10001] 0 1100 4553 8336 12371 13377 13903 19122 23878 28668 ...
  ..@ Dim     : int [1:2] 33391 10000
  ..@ Dimnames:List of 2
  .. ..$ : chr [1:33391] "t33417aep-AAPT1-ARATH" "t33418aep-ZBED8-BOVIN" "t37645aep-CP4V2-MOUSE" "t31628aep-TSP1-BOVIN" ...
  .. ..$ : chr [1:10000] "02-P1_GTCAGCGCTGTG" "02-P1_TGAGAGTGAGCC" "06-FM_GCCGTACAGCAT" "11-BU_CGTCTACATATA" ...
  ..@ x       : num [1:22443960] 1 1 1 1 2 1 1 1 2 10 ...
  ..@ factors : list()
}
\details{
The Hydra scRNA-seq dataset contains an extensive molecular description of the
adult Hydra polyp
}
\source{
https://www.dropbox.com/sh/fwx553vie2eck5v/AACZEjWEvpag-wvx9MWhblzXa?dl=0
}
\references{
In the file Hydra_Seurat_Whole_Transcriptome.rds from the link above
}
\examples{
data(input_counts)
## maybe str(input_counts) ; plot(input_counts) ...
}
\keyword{datasets}
